# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Health Plan Factory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to returnTo on success or /api/login on failure.
...
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
...
  /employer/dashboard:
    get:
      tags: [employer]
      operationId: getEmployerDashboard
      summary: Get employer dashboard stats including spend, utilization, and top modalities
      description: >
        Returns aggregate employer wellness programme statistics. Utilization
        and spend metrics are suppressed (null) when fewer than 5 members are
        enrolled to protect individual privacy (k-anonymity floor).
      responses:
        '200':
          description: Employer dashboard data.
          content:
            application/json:
              schema:
                type: object
                required: [employer, stats, topModalities, monthlySpend]
                properties:
                  employer:
                    $ref: '#/components/schemas/Employer'
                  stats:
                    type: object
                    required: [totalEnrolled, totalBudgetCents, monthlyInvoiceCents]
                    properties:
                      totalEnrolled:
                        type: integer
                      totalBudgetCents:
                        type: integer
                      totalSpentCents:
                        type: integer
                        nullable: true
                        description: Suppressed (null) when cohort < 5 members
                      utilizationPct:
                        type: integer
                        nullable: true
                        description: Suppressed (null) when cohort < 5 members
                      avgWellnessScore:
                        type: integer
                        nullable: true
                        description: Suppressed (null) when cohort < 5 members
                      monthlyInvoiceCents:
                        type: number
                      privacySuppressed:
                        type: boolean
                  topModalities:
                    type: array
                    items:
                      type: object
                      required: [modalityId, sessionCount]
                      properties:
                        modalityId:
                          type: string
                        sessionCount:
                          type: integer
                  monthlySpend:
                    type: array
                    description: 6-month historical spend derived from actual employer-covered session costs
                    items:
                      type: object
                      required: [month, totalCents]
                      properties:
                        month:
                          type: string
                          description: "YYYY-MM"
                        totalCents:
                          type: integer
        '404':
          description: No employer account found.
...
components:
...
  schemas:
...
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    Employer:
      type: object
      required: [id, companyName, status, inviteCode]
      properties:
        id:
          type: string
        companyName:
          type: string
        status:
          type: string
          enum: [pending, active, canceled]
        inviteCode:
          type: string
        numberOfEmployees:
          type: integer
        stipendPerEmployee:
          type: integer
          description: Monthly stipend per employee in cents
        platformFeePercent:
          type: number
        createdAt:
          type: string
          format: date-time
...